[DEV-6264] Split events_controller module
