+++ server/athenian/api/controllers/events_controller/__init__.py
@@ -1,0 +1,9 @@
+from athenian.api.controllers.events_controller.clear_precomputed_events import (
+    clear_precomputed_events,
+)
+from athenian.api.controllers.events_controller.deployments import notify_deployments
+from athenian.api.controllers.events_controller.deployments_labels import (
+    get_deployment_labels,
+    modify_deployment_labels,
+)
+from athenian.api.controllers.events_controller.releases import notify_releases
